+++ builder/vmware/guest_ip.go
@@ -1,7 +1,7 @@
 package vmware
 
 import (
-	"fmt"
+	"errors"
 	"io/ioutil"
 	"os"
 	"regexp"
@@ -17,6 +17,9 @@ type GuestIPFinder interface {
 // DHCPLeaseGuestLookup looks up the IP address of a guest using DHCP
 // lease information from the VMware network devices.
 type DHCPLeaseGuestLookup struct {
+	// Driver that is being used (to find leases path)
+	Driver Driver
+
 	// Device that the guest is connected to.
 	Device string
 
@@ -25,7 +28,7 @@ type DHCPLeaseGuestLookup struct {
 }
 
 func (f *DHCPLeaseGuestLookup) GuestIP() (string, error) {
-	fh, err := os.Open(fmt.Sprintf("/var/db/vmware/vmnet-dhcpd-%s.leases", f.Device))
+	fh, err := os.Open(f.Driver.DhcpLeasesPath(f.Device))
 	if err != nil {
 		return "", err
 	}
@@ -66,6 +69,10 @@ func (f *DHCPLeaseGuestLookup) GuestIP() (string, error) {
 			curIp = lastIp
 			curLeaseEnd = lastLeaseEnd
 		}
+	}
+
+	if curIp == "" {
+		return "", errors.New("IP not found for MAC in DHCP leases")
 	}
 
 	return curIp, nil